requirements.txt
fastapi
uvicorn
aiofiles
sqlalchemy
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pymysql
psycopg2-binary